Implement api/cryptos routes, update schema
What:

1) Implemented /cryptos routes
2) Changed quantity data type to decimal from int in crypto tables
3) Added ICryptoInfo type

Why:

1) Needed
2) crypto won't always be purchased in even numbered amounts
3) Needed

+++ server/src/db/schema/09_Portfolios_Cryptos.sql
@@ -4,6 +4,6 @@ CREATE TABLE portfolios_cryptos (
   id serial PRIMARY KEY NOT NULL,
   crypto_id integer NOT NULL REFERENCES cryptos (id) ON DELETE CASCADE,
   portfolio_id integer NOT NULL REFERENCES portfolios (id) ON DELETE CASCADE,
-  quantity integer NOT NULL,
+  quantity decimal NOT NULL,
   unique (portfolio_id, crypto_id)
 );
